Add files via upload
Measurement and monitoring of CO and NOx emissions using periodic measurement methods and CEMS has the potential for data inaccuracies. The inaccuracy of the data can be caused by the failure of measuring devices or sensors to measure variable values at the test location. Therefore, to ensure the accuracy of the data, PEMS is needed to predict the amount of CO and NOx emissions based on independent process variables. PEMS prediction results will be compared with direct measurement data in the field by looking for error values. A small error value between the PEMS and CEMS methods or periodic measurements indicates that the sensor has no problems in measurement. PEMS will use one of the methods in machine learning to create a predictive model for CO and NOx emissions.